add cmd and move some code into internal
