+++ src/pkg/cli/internal/cmd/lifecycle/evict.go
@@ -4,7 +4,7 @@ import (
 	"github.com/spf13/cobra"
 	"k8s.io/client-go/tools/clientcmd"
 
-	"github.com/deckhouse/virtualization/src/pkg/cli/templates"
+	"github.com/deckhouse/virtualization/src/pkg/cli/internal/templates"
 )
 
 func NewEvictCommand(clientConfig clientcmd.ClientConfig) *cobra.Command {
